pyproject.toml
[project]
name = "flaskproject"
version = "0.1.0"
description = "flask project"
requires-python = ">=3.13"
dependencies = [
    "email-validator==2.2.0",
    "flask==3.1.1",
    "flask-bcrypt>=1.0.1",
    "flask-debugtoolbar>=0.16.0",
    "flask-limiter>=3.12",
    "flask-login>=0.6.3",
    "flask-mail>=0.10.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "pyjwt>=2.10.1",
    "python-dotenv==1.1.0",
    "wtforms>=3.2.1",
]
